requirements.txt
boto3==1.26.142
botocore==1.29.142
Django==4.0.3
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
django-cors-headers